(semi-) working panel using Clutter. No taskbar.
The workspace indicators kind of work, although they're a massive hack at the moment. I need to set up a MVC-type structure so everything stays up-to-date.

The taskbar is currently a lower priority than restructuring the code. It's quite likely filenames will move around, and things will be reorganised to be easier to maintain. Wish me luck, me!