Working, clean panel with pager.
Still no tasklist, but the pager is working, and it is nice~.
The PagerModel might need a little cleanup to improve its API, but in general, it's OK.

I switched from plain Clutter to GtkClutter, so the WM hints are set up correctly. It has a few extra dependencies (Gtk, GtkClutter), but that shouldn't really be an issue. They're installed on my system most always anyway...