Working (mostly) taskbar!
Windows can be minimised, raised, and closed with the taskbar. It doesn't handle opening new windows yet; this is a problem with the model/view that I've constructed so far. I might have to look into Clutter's built-in mechanisms for such things. However, it's on the way to being a usable + useful panel.