_NET_WM_STRUT support.
We are now ready for actual use! STRUT support requires xprop to set things, but as we're already requiring Wnck, Clutter, etc, that's fine. I can finally use it to replace tint2! Still not ready for general use, as it's hardcoded to use 1366x16 as the size, but oh well...