Panel "order" works now.
The order of things is properly set now. If you wanted, you could have the panel on the right-hand side of the screen, or a clock in the centre like Gnome3. It's pretty cool, anyway.

I need to have a look at writing up some proper documentation. The default config should probably have examples for every option, if commented out.